move packages: util.kt -> "parser.antlr"
