+++ src/test/kotlin/io/github/durun/nitron/app/preparse/ExtractorTest.kt
@@ -1,10 +1,10 @@
 package io.github.durun.nitron.app.preparse
 
 import io.github.durun.nitron.core.MD5
-import io.github.durun.nitron.core.ast.type.nodeTypePoolOf
 import io.github.durun.nitron.core.config.loader.NitronConfigLoader
 import io.github.durun.nitron.core.parser.ParserStore
 import io.github.durun.nitron.core.parser.antlr.AstBuildVisitor
+import io.github.durun.nitron.core.parser.antlr.nodeTypePoolOf
 import io.github.durun.nitron.inout.database.MemoryDatabase
 import io.github.durun.nitron.inout.model.preparse.*
 import io.kotest.core.spec.style.FreeSpec
