+++ src/test/kotlin/io/github/durun/nitron/core/ast/processors/AstNormalizerTest.kt
@@ -1,7 +1,7 @@
 package io.github.durun.nitron.core.ast.processors
 
 import io.github.durun.nitron.core.ast.path.AstPath
-import io.github.durun.nitron.core.ast.type.nodeTypePoolOf
+import io.github.durun.nitron.core.parser.antlr.nodeTypePoolOf
 import io.github.durun.nitron.core.config.loader.LangConfigLoader
 import io.github.durun.nitron.core.parser.ParserStore
 import io.github.durun.nitron.core.parser.antlr.AstBuildVisitor
