+++ src/test/kotlin/io/github/durun/nitron/core/ast/processors/AstSplitterTest.kt
@@ -1,6 +1,6 @@
 package io.github.durun.nitron.core.ast.processors
 
-import io.github.durun.nitron.core.ast.type.nodeTypePoolOf
+import io.github.durun.nitron.core.parser.antlr.nodeTypePoolOf
 import io.github.durun.nitron.core.config.loader.LangConfigLoader
 import io.github.durun.nitron.core.parser.ParserStore
 import io.github.durun.nitron.core.parser.antlr.AstBuildVisitor
